requirements.txt
# Web Framework
Flask==3.0.0
gunicorn==21.2.0

# Task Queue & Message Broker
celery==5.3.4
redis==5.0.1

# Database & ORM
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23
alembic==1.13.1

# Data Validation
pydantic==2.5.0

# Social Media Integration
tweepy==4.14.0

# Machine Learning
transformers==4.36.2
torch==2.1.2
numpy==1.24.3

# Environment & Configuration
python-dotenv==1.0.0

# Development & Testing for future 
pytest==7.4.3
pytest-flask==1.3.0
black==23.12.1
flake8==6.1.0
